This is a big bad bundle commit containing code by Brother John and Mike.
The main change is a completely new, modular architecture for the filters. The "filters" subfolder contains the major parts including already ported filter classes.

Other major changes
-------------------
* Added: ptInfo, a very basic global logging/error handling system. To be expanded.
* Added: ptWidget, the new-style base class for custom Photivo widgets. The traditional ptCheck, ptChoice and ptInput remain in use but are based on ptWidget to make them work with new-style filters.
* Added: ptToolBox, the new-style foldable box for the GUI of a filter. Is gradually replacing ptGroupBox.
* Changed: the central data structure in ptImage from C-style memory management to STL containers. Work in progress.

Note that ptCreateSomeCurves.cpp is currently broken. The ptCreateCurvesProject subproject cannot be built. It is unclear if/how that subproject is still needed since the way how curves are handled and stored changed quite a bit in the wake of the new filter architecture. For the end user nothing changes; ptCreateCurvesProject always has been a developer only tool.

New build requirements
----------------------
The requirements to build Photivo have changed, mainly because the new code started using C++11 - and that is a trend that will continue heavily. At the moment at least
* Qt 4.7 and
* GCC 4.6
are needed. A move to Qt 4.8 and GCC 4.7 is likely in the foreseeable future.

+++ Sources/rawtherapee/cfa_line_dn.c
@@ -111,8 +111,8 @@ void CLASS cfa_linedn(float noise)
   for (int top=0; top < height-16; top += TS-32)
     for (int left=0; left < width-16; left += TS-32) {
 
-      int bottom = MIN( top+TS,height);
-      int right  = MIN(left+TS, width);
+      int bottom = MIN( top+TS, (int32_t)height);
+      int right  = MIN(left+TS, (int32_t)width);
       int numrows = bottom - top;
       int numcols = right - left;
       int indx1;
